backend/requirements.txt
dj-database-url
Django
django-cors-headers
djangorestframework
djangorestframework-simplejwt